% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hipo.R
\name{hipo}
\alias{hipo}
\title{Heritability Informed Power Optimization (HIPO)}
\usage{
hipo(sumstats, out.path, maf.thr = 0.05, HIPOD.num = 2,
  ldsc.path = "~/documents/software/ldsc", python.path = NULL,
  ldscore.path = file.path(ldsc.path, "eur_w_ld_chr/"), mergeallele = TRUE,
  truncate = NULL)
}
\arguments{
\item{sumstats}{A list of K containing summary statistics for K traits. Each element of the list contains 9 columns：rsid, A1 (effect allele), A2, N (sample size), z (z-statistic), pval, chr (chromosome number), bp (base pair; physical position), freqA1 (allele frequency of A1).}

\item{maf.thr}{MAF threshold in SNP QC. SNPs with MAF < maf.thr are removed. Default 0.05.}

\item{HIPOD.num}{Number of HIPO components to calculate z stats and p-value for.}

\item{ldsc.path}{The directory of the LDSC software.}

\item{python.path}{The path of Python, if need to use a version other than the default one.}

\item{ldscore.path}{The path containing the LDSC LD score files. Default to be file.path(ldsc.path,"eur_w_ld_chr/").}

\item{mergeallele}{Whether to merge allele to HapMap 3 SNPs when using LDSC. Default to be TRUE.}

\item{truncate}{If NULL, use the full coherit.mat and ldscint.mat; if a decimal, truncate the eigenvectors of ldscint.mat with eigenvalues < truncate*(max eigenvalue of ldscint.mat). Recommended truncate = 0.05.}
}
\description{
Heritability Informed Power Optimization (HIPO)
}

