% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hipoeigen.R
\name{hipoeigen}
\alias{hipoeigen}
\title{HIPO eigendecomposition}
\usage{
hipoeigen(sumstats.all, traitvec, coherit.mat, ldscint.mat, HIPOD.num = 2,
  truncate = NULL)
}
\arguments{
\item{sumstats.all}{a list of K that contains summary statistics for K traits}

\item{traitvec}{Vector of trait names, second output of preprocess().}

\item{coherit.mat}{Genetic covariance matrix.}

\item{ldscint.mat}{LD score regression intercept matrix.}

\item{HIPOD.num}{Number of HIPO components to calculate z stats and p-value for.}

\item{truncate}{If NULL, use the full coherit.mat and ldscint.mat; if a decimal, truncate the eigenvectors of ldscint.mat with eigenvalues < truncate*(max eigenvalue of ldscint.mat). Recommended truncate = 0.05.}
}
\value{
A list that contains

sumstats.all = A data.frame containing the merged individual trait summary statistics and the z statistic (z.HIPODx) and p-value (pval.HIPODx) of HIPO components.

\item{coherit.mat}{Heritability-coheritability matrix.}
\item{ldscint.mat}{Matrix of LD score regression intercepts.}
\item{eigenvalue}{Eigenvalues from HIPO eigendecomposition. Proportional to the average non-centrality parameter.}
\item{HIPOD.mat}{A matrix of which the columns are the HIPO components.}
}
\description{
This function performs HIPO eigendecomposition and calculate the z statistics and p-values of HIPO components.
}

